backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-multipart==0.0.20
httpx==0.28.1
pdfplumber==0.11.5
python-docx==1.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.4
pytest-asyncio==0.25.0
